+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fix a bug in the interactive mode.
+- Fix a bug in the seq2seq model's output.
+
+## [v2.1.0] - 2023-06-02
 ### Added
 - Support Python 3.11.
 - Support NN-based sentence segmentation.
@@ -21,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   model_size: base
   device: cpu
   num_workers: 0
+  torch_compile: false
   typo_batch_size: 1
   senter_batch_size: 1
   seq2seq_batch_size: 1
@@ -156,7 +162,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - Remove an unnecessary dependency, `fugashi`.
 
-[Unreleased]: https://github.com/ku-nlp/kwja/compare/v2.0.0...HEAD
+[Unreleased]: https://github.com/ku-nlp/kwja/compare/v2.1.0...HEAD
+[2.1.0]: https://github.com/ku-nlp/kwja/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/ku-nlp/kwja/compare/v1.4.2...v2.0.0
 [1.4.2]: https://github.com/ku-nlp/kwja/compare/v1.4.1...v1.4.2
 [1.4.1]: https://github.com/ku-nlp/kwja/compare/v1.4.0...v1.4.1